Replace floating_ui.js over index.js files
